requirements.txt
aiohttp
aiohttp-retry==2.8.3
aiosignal==1.3.1
alembic==1.13.2
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
docx==0.2.4
dominate==2.9.1
ffmpeg==1.4
Flask==2.3.2
Flask-Admin==1.6.1
Flask-Bootstrap==3.3.7.1
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
frozenlist==1.4.1
google-api-core==2.19.1
google-auth==2.32.0
google-cloud-speech==2.27.0
googleapis-common-protos==1.63.2
greenlet==3.1.0
grpcio==1.65.2
grpcio-status==1.65.2
gunicorn==22.0.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
lxml==5.2.2
Mako==1.3.5
MarkupSafe==2.1.5
multidict==6.0.5
numpy==1.26.4
pandas==2.2.2
pillow==10.3.0
prettytable==3.10.0
proto-plus==1.24.0
protobuf==5.27.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydub==0.25.1
PyJWT==2.8.0
pyperclip==1.8.2
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
rsa==4.9
six==1.16.0
soupsieve==2.6
SpeechRecognition==3.10.4
SQLAlchemy==2.0.35
twilio==9.2.2
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
visitor==0.1.3
wcwidth==0.2.13
Werkzeug==3.0.4
WTForms==3.1.2
yarl==1.9.4
supabase~=2.7.4